+++ backend/src/db/impls/mod.rs
@@ -1,0 +1,6 @@
+mod nlp_analysis;
+mod nlp_metadata;
+mod report;
+mod report_analyses_map;
+mod report_metadata;
+mod user;
