+++ backend/src/db/impls/user.rs
@@ -1,0 +1,70 @@
+use crate::api::auth::google::User;
+use crate::db::db_stored::DbStoredInner;
+use crate::db::pagination::DbPagination;
+use axum::async_trait;
+use sqlx::{Error, Postgres, Transaction};
+
+#[async_trait]
+impl DbStoredInner for User {
+    async fn inner_save(&self, db: &mut Transaction<Postgres>) -> Result<(), Error> {
+        sqlx::query!(
+            r#"
+            INSERT INTO users (
+            google_sub, name, given_name, family_name, picture, email, email_verified
+            )
+            VALUES ($1, $2, $3, $4, $5, $6, $7)
+            ON CONFLICT (google_sub) DO UPDATE
+            SET name = $2, given_name = $3, family_name = $4, picture = $5, email = $6, email_verified = $7
+            "#,
+            self.id,
+            self.name,
+            self.given_name,
+            self.family_name,
+            self.picture,
+            self.email,
+            self.email_verified
+        )
+        .execute(&mut **db)
+        .await?;
+        Ok(())
+    }
+    async fn inner_update(&self, db: &mut Transaction<Postgres>) -> Result<(), Error> {
+        unimplemented!()
+    }
+    async fn inner_delete(&self, db: &mut Transaction<Postgres>) -> Result<(), Error> {
+        sqlx::query!(
+            r#"
+            DELETE FROM users WHERE google_sub = $1
+            "#,
+            self.id
+        )
+        .execute(&mut **db)
+        .await?;
+        Ok(())
+    }
+
+    async fn inner_get_by_id(
+        id: &str,
+        db: &mut Transaction<Postgres>,
+    ) -> Result<Option<Self>, Error> {
+        let user = sqlx::query_as!(
+            User,
+            r#"
+            SELECT google_sub as "id: String", name, given_name, family_name, picture, email, email_verified
+            FROM users
+            WHERE google_sub = $1
+            "#,
+            id
+        )
+        .fetch_optional(&mut **db)
+        .await?;
+        Ok(user)
+    }
+
+    async fn inner_get_all(
+        pagination: DbPagination,
+        db: &mut Transaction<Postgres>,
+    ) -> Result<Vec<Self>, Error> {
+        todo!()
+    }
+}
