+++ backend/src/db/mod.rs
@@ -1,3 +1,5 @@
-pub mod db_client;
+pub(crate) mod db_client;
 pub(crate) mod db_stored;
+mod impls;
 mod model;
+mod pagination;
