+++ backend/src/db/pagination.rs
@@ -1,0 +1,4 @@
+pub struct DbPagination {
+    page: u32,
+    per_page: u32,
+}
